requirements.txt
arrow==0.14.6
certifi==2019.6.16
chardet==3.0.4
dateparser==0.7.1
get==2019.4.13
humanize==0.5.1
idna==2.8
loguru==0.3.2
maya==0.6.1
numpy==1.17.1
pandas==0.25.1
pendulum==2.0.5
polyline==1.4.0
post==2019.4.13
public==2019.4.13
pyarrow==0.14.1
python-dateutil==2.8.0
pytz==2019.2
pytzdata==2019.2
PyYAML==5.1.2
query-string==2019.4.13
regex==2019.8.19
requests==2.22.0
six==1.12.0
snaptime==0.2.4
stravaio==0.0.9
tzlocal==2.0.0
units==0.7
urllib3==1.25.3
